Move resolver tests out to top-level
